Consolidating the pathCount tests, removing missing dependencies.
