+++ lib/pathCount.js
@@ -30,20 +30,20 @@ function getRangeSize(range) {
     var length = range.length;
 
     if (to != null) {
-        if (isNaN(to) || parseInt(to, 10) != to) { // eslint-disable-line eqeqeq
+        if (isNaN(to) || parseInt(to, 10) !== to) {
             throw new Error("Invalid range, 'to' is not an integer: " + JSON.stringify(range));
         }
         var from = range.from || 0;
-        if (isNaN(from) || parseInt(from, 10) != from) { // eslint-disable-line eqeqeq
+        if (isNaN(from) || parseInt(from, 10) !== from) {
             throw new Error("Invalid range, 'from' is not an integer: " + JSON.stringify(range));
         }
         if (from <= to) {
             return (to - from) + 1;
         } else {
-            throw new Error("Invalid range, 'to' smaller than 'from': " + JSON.stringify(range));
+            return 0;
         }
     } else if (length != null) {
-        if (isNaN(length) || parseInt(length, 10) != length) { // eslint-disable-line eqeqeq
+        if (isNaN(length) || parseInt(length, 10) !== length) {
             throw new Error("Invalid range, 'length' is not an integer: " + JSON.stringify(range));
         } else {
             return length;
